Add redirects to login
Also refactors the App component to render a BrowserRouter, and updates the
App component tests to check the redirect. We might get bitten by this later if
our app starts to include complex redirecting logic -- I know that there is a
MemoryRouter in the react-router library that seems to be intended for use with
tests.

Also adds a failing e2e test for a successful login flow, to give us a place to
start next time!

Signed-off-by: Diego Moura <diegoo.demoura@gmail.com>
Co-authored-by: Jamie Klassen <jklassen@vmware.com>